Enhance and correct rendering and image output
Add dynamic datestr in the image file using time. Format is YYMMDDHHMMSS.

Set non-escaping pixels to black (0) in the intensity mapping functions.

Add color image output in the pybrot module. Automatically detect colour
or greyscale arrays in pybrot's RawImage. Colour arrays should make use
of the boxed pixel format for correct detection.

Add pretty colormaps via matplotlib.

Move escape check before iteration and loop break code in mandel.cpp.
This affects the points starting outside a radius 2 circle, which escape
immediately with an iteration count of 0.

Move intensity mapping functions ahead of code to allow definition of a
custom mapping lambda function in the configuration variables. Change
intensity mapping functions to work with an input range from 0.0 to 1.0
and output in a range from 0.0 to 1.0. This allows chaining multiple
mapping functions as needed, and allows a normalization of the input as
needed.

Change color_log mapping function to conform with
https://homepages.inf.ed.ac.uk/rbf/HIPR2/pixlog.htm. The new formula
does not need checks for log(0).
The formula
log(1 + I*f, 1 + f)
where I is the sample intensity also adds a factor f to control the
logarithmic curve dropoff.

Add option and function to linearly normalize raw data to max and min
values before passing it to the intensity mapping functions.

Add color_sigmoid mapping function. This can be useful for contrast
enhancement. The basic sigmoid curve is calculated with
sig = 1.0 / (1.0 + e^(-(I-0.5)/a))
where I is the sample intensity and a is a value adjusting the sharpness
of the curve dropoff.
Using
(sig - 0.5) * 1.0 / (1.0 - 2.0 / (1.0 + e^(0.5/a))) + 0.5
The sigmoid is renormalized to deliver a full 0.0-1.0 output range.

Flip calculations in mandel.cpp to start rendering at the upper edge of
the image and conform with the pybrot RawImage/pypng output orientation.

Add and comment out -O3 optimization in CMakeLists.txt.